merge

+++ x-pack/plugin/sql/src/test/java/org/elasticsearch/xpack/sql/expression/function/scalar/datetime/DateTimeToCharProcessorTests.java
@@ -49,7 +49,7 @@ import static org.elasticsearch.xpack.sql.expression.function.scalar.datetime.To
  *     </li>
  * </ol>
  *
- * In case you need to mute any of the tests, mute all tests by adding {@link org.apache.lucene.util.LuceneTestCase.AwaitsFix}
+ * In case you need to mute any of the tests, mute all tests by adding {@link org.apache.lucene.tests.util.LuceneTestCase.AwaitsFix}
  * on the class level.
  */
 public class DateTimeToCharProcessorTests extends ESTestCase {
@@ -163,7 +163,7 @@ public class DateTimeToCharProcessorTests extends ESTestCase {
         int fractions = b.remainder(BigDecimal.ONE).movePointRight(9).intValueExact();
         int adjustment = 0;
         if (fractions < 0) {
-            fractions += 1e9;
+            fractions += (int) 1e9;
             adjustment = -1;
         }
         return dateTime((seconds + adjustment) * 1000).withNano(fractions);
